Fix timings, start building map/reduce scheduling for vehicles

+++ src/main/scala/edu/rpi/cs/nsl/spindle/vehicle/Types.scala
@@ -2,5 +2,5 @@ package edu.rpi.cs.nsl.spindle.vehicle
 
 object Types {
   type NodeId = Int // UID for vehicle in database
-  type Timestamp = Double
+  type Timestamp = Long
 }
